requirements.txt
streamlit

fastapi

uvicorn

langchain

langchain-community

chromadb

sentence-transformers

pypdf

python-dotenv

ollama

typing-extensions

